Re-assigned eval board button actions.
Added buttons to user interface on either size of the main viewing
area to simulate the physical buttons that will be present for the
finished product.  The buttons on the left side of the graphics
correspond to the column of buttons closest to the eval board
LCD screen.  The buttons on the right side correspond to the column
of buttons closest to the edge of the eval board.